+++ cursor/src/main/kotlin/cursors/effects/Show.kt
@@ -7,6 +7,7 @@ import cursors.io.right
 import cursors.io.scala2s
 import vec.macros.size
 import vec.macros.toList
+import kotlin.math.min
 
 /*simple printout macro*/
 fun Cursor.show(range: IntProgression = 0 until size) {
@@ -19,3 +20,4 @@ fun Cursor.showValues(range: IntProgression) {
         println((this at it).left.toList())
     }
 }
+fun Cursor.head ( last:Int)=show(0 until (min(last,size)))
